fix(prefab): use latest Unity XR Toolkit components
The Pico SDK has a hard dependency on the Unity XR Toolkit but on a
very old 0.9.4 preview version, which is now vastly outdated. This
ensures the latest XR Toolkit version is installed and updates the
prefab to use the XR Origin component instead of the XR Rig component
as that is now deprecated.